Add score module, max score reachable set, tha way the game displays de score and the font resize depending of the amoun of characters that the score have.
Add conditions to not move the blocks when the game is paused, when the game is waiting for the user input to restart the game and not to continue scoring points when the game is paused.

Refactor the way the font size is set and update the values of some methods that were receiveing a dobule variable instead a floar variable.